+++ src/main/java/net/minestom/arena/group/GroupManager.java
@@ -1,6 +1,6 @@
 package net.minestom.arena.group;
 
-import net.minestom.arena.LobbySidebarDisplay;
+import net.minestom.arena.lobby.LobbySidebarDisplay;
 import net.minestom.arena.Messenger;
 import net.minestom.server.entity.Player;
 import org.jetbrains.annotations.NotNull;
